CompactFilterBehavior: Extract `FilterSynchronizationState` to a standalone file
